+++ package.json
@@ -67,7 +67,7 @@
     "@tanstack/react-table": "8.21.2",
     "@tanstack/react-virtual": "^3.11.2",
     "@vercel/og": "0.6.5",
-    "@vertigo-amm/vertigo-sdk": "0.1.7",
+    "@vertigo-amm/vertigo-sdk": "0.1.8",
     "bs58": "6.0.0",
     "class-variance-authority": "^0.7.1",
     "classnames": "^2.5.1",
@@ -98,6 +98,7 @@
     "tailwind-merge": "3.0.1",
     "tailwind-scrollbar-hide": "2.0.0",
     "tailwindcss-animate": "1.0.7",
+    "yargs": "17.7.2",
     "zod": "3.24.2",
     "zustand": "^5.0.2"
   },
